+++ core/server/data/schema/clients/mysql.js
