+++ core/server/data/schema/clients/sqlite3.js
