+++ core/server/data/schema/default-settings.json
